feat: Payment 도메인 Price 필드 자료형 수정 및 도메인 변화로 인한 변경

+++ persistence/persist-payment/src/main/java/store/nightmarket/persistence/persistpayment/entity/valueobject/PriceEntity.java
@@ -1,7 +1,5 @@
 package store.nightmarket.persistence.persistpayment.entity.valueobject;
 
-import java.math.BigDecimal;
-
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -16,6 +14,6 @@ import lombok.NoArgsConstructor;
 @EqualsAndHashCode
 public class PriceEntity {
 
-	private BigDecimal price;
+	private long price;
 
 }
